Add Refresh interval to settings
# Conflicts:
#	PRChecker/PRChecker.xcodeproj/project.pbxproj
