+++ PRChecker/Shared/Unique Views/SettingsView.swift
@@ -18,19 +18,11 @@ struct SettingsView: View {
                 
                 Divider()
                 
-                List {
-                    ForEach(settingsViewModel.userList, id: \.self) { user in
-                        WatchedUserView(username: user) { settingsViewModel.remove($0) }
-                    }
-                    .onDelete { indexSet in
-                        settingsViewModel.remove(indexSet)
-                    }
-                }
-                .frame(height: 300)
+                RefreshSettingsView(settingsViewModel: settingsViewModel)
                 
-                Spacer()
-                AddUserView() { settingsViewModel.addUser($0) }
-                .padding([.leading, .trailing])
+                Divider()
+                
+                WatchedListView(settingsViewModel: settingsViewModel)
             }
             .padding()
         }
@@ -41,44 +33,5 @@ struct SettingsView: View {
 struct SettingsView_Previews: PreviewProvider {
     static var previews: some View {
         SettingsView()
-    }
-}
-
-struct WatchedUserView: View {
-    var username: String
-    var onRemove: ((String) -> Void)?
-    
-    var body: some View {
-        HStack {
-            Image(systemName: "person.fill")
-                .font(.title)
-            Text(username)
-                .font(.title)
-            Spacer()
-            Button {
-                onRemove?(username)
-            } label: {
-                Image(systemName: "x.circle.fill")
-                    .foregroundColor(.red)
-            }
-            .buttonStyle(PlainButtonStyle())
-        }
-        .padding([.leading, .trailing, .top])
-    }
-}
-
-struct AddUserView: View {
-    @State var newUsername = ""
-    var onAdd: ((String) -> Void)?
-    
-    var body: some View {
-        HStack {
-            TextField(LocalizedStringKey("Username"), text: $newUsername)
-            Button("Add") {
-                guard !newUsername.isEmpty else { return }
-                onAdd?(newUsername)
-                newUsername = ""
-            }
-        }
     }
 }
